requirements.txt
Bio==1.4.0
dgl==0.9.1
dgl_cu111==0.8.2
h5py==2.10.0
joblib==0.17.0
numpy==1.22.4
pandas==1.4.2
pytz==2020.1
PyYAML==6.0
scikit_learn==0.23.2
torch==1.11.0+cu113
tqdm==4.50.2
wget==3.2
